pyproject.toml
[project]
name = "movie-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "gradio>=6.11.0",
    "keybert>=0.9.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "tqdm>=4.67.3",
    "transformers>=4.55.0",
    "youtube-search>=2.2.0",
]

[project.scripts]
moviemate = "src.main:main"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=8.3.3",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
only-include = ["src"]
